+++ http4k-core/src/main/kotlin/org/http4k/server/PolyHandler.kt
@@ -1,10 +1,14 @@
 package org.http4k.server
 
 import org.http4k.core.HttpHandler
+import org.http4k.sse.SseHandler
 import org.http4k.websocket.WsHandler
 
 /**
- * A PolyHandler represents the combined routing logic of an Http handler and a Websocket handler.
- * ws:// and http:// protocol calls are passed relevantly.
+ * A PolyHandler represents the combined routing logic of an multiple protocol handlers
  */
-class PolyHandler(val http: HttpHandler?, internal val ws: WsHandler?)
+class PolyHandler @JvmOverloads constructor(
+    val http: HttpHandler? = null,
+    val ws: WsHandler? = null,
+    val sse: SseHandler? = null
+)