+++ http4k-core/src/main/kotlin/org/http4k/websocket/deprecatedCore.kt
@@ -1,0 +1,4 @@
+package org.http4k.websocket
+
+@Deprecated("moved to org.http4k.server", ReplaceWith("org.http4k.server.PolyHandler"))
+typealias PolyHandler = org.http4k.server.PolyHandler