+++ http4k-server/jetty/src/main/kotlin/org/http4k/server/jetty.kt
@@ -20,18 +20,21 @@ import org.eclipse.jetty.websocket.core.server.WebSocketUpgradeHandler
 import org.http4k.core.HttpHandler
 import org.http4k.servlet.asHttp4kRequest
 import org.http4k.servlet.asServlet
+import org.http4k.sse.SseHandler
 import org.http4k.websocket.WsHandler
+import java.lang.UnsupportedOperationException
 
 
-class Jetty(private val port: Int, private val server: Server) : WsServerConfig {
+class Jetty(private val port: Int, private val server: Server) : PolyServerConfig {
     constructor(port: Int = 8000) : this(port, http(port))
     constructor(port: Int, vararg inConnectors: ConnectorBuilder) : this(port, Server().apply {
         inConnectors.forEach { addConnector(it(this)) }
     })
 
-    override fun toServer(httpHandler: HttpHandler?, wsHandler: WsHandler?): Http4kServer {
-        httpHandler?.let { server.insertHandler(httpHandler.toJettyHandler()) }
-        wsHandler?.let {
+    override fun toServer(http: HttpHandler?, ws: WsHandler?, sse: SseHandler?): Http4kServer {
+        if(sse != null) throw UnsupportedOperationException("Jetty does not support sse")
+        http?.let { server.insertHandler(http.toJettyHandler()) }
+        ws?.let {
             server.insertHandler(
                 WebSocketUpgradeHandler(
                     WebSocketComponents()).apply {