+++ http4k-server/undertow/src/main/kotlin/org/http4k/server/Undertow.kt
@@ -1,32 +1,39 @@
 package org.http4k.server
 
 import io.undertow.Handlers.predicate
+import io.undertow.Handlers.serverSentEvents
 import io.undertow.Handlers.websocket
 import io.undertow.Undertow
 import io.undertow.UndertowOptions.ENABLE_HTTP2
 import io.undertow.server.handlers.BlockingHandler
 import org.http4k.core.HttpHandler
+import org.http4k.sse.SseHandler
 import org.http4k.websocket.WsHandler
 import java.net.InetSocketAddress
 
-data class Undertow(val port: Int = 8000, val enableHttp2: Boolean) : WsServerConfig {
+data class Undertow(val port: Int = 8000, val enableHttp2: Boolean) : PolyServerConfig {
     constructor(port: Int = 8000) : this(port, false)
 
-    override fun toServer(httpHandler: HttpHandler?, wsHandler: WsHandler?): Http4kServer {
-        val http = httpHandler?.let(::HttpUndertowHandler)?.let(::BlockingHandler)
-        val ws = wsHandler?.let { websocket(Http4kWebSocketCallback(it)) }
+    override fun toServer(http: HttpHandler?, ws: WsHandler?, sse: SseHandler?): Http4kServer {
+        val httpHandler = http?.let(::Http4kUndertowHttpHandler)?.let(::BlockingHandler)
+        val wsCallback = ws?.let { websocket(Http4kWebSocketCallback(it)) }
+        val sseCallback = sse?.let { serverSentEvents(Http4kSseCallback(sse)) }
 
-        val handler = when {
-            http != null && ws != null -> predicate(requiresWebSocketUpgrade(), ws, http)
-            ws != null -> ws
-            else -> http
+        val handlerWithWs = when {
+            httpHandler != null && wsCallback != null -> predicate(requiresWebSocketUpgrade(), wsCallback, httpHandler)
+            wsCallback != null -> wsCallback
+            else -> httpHandler
         }
+
+        val handlerWithSse = sseCallback
+            ?.let { predicate(hasEventStreamContentType(), sseCallback, handlerWithWs) }
+            ?: handlerWithWs
 
         return object : Http4kServer {
             val server = Undertow.builder()
                 .addHttpListener(port, "0.0.0.0")
                 .setServerOption(ENABLE_HTTP2, enableHttp2)
-                .setHandler(handler).build()
+                .setHandler(handlerWithSse).build()
 
             override fun start() = apply { server.start() }